Update comments/tests wrt OS 3.20 OSWORD &0F
Ken has investigated this on real hardware and I think the comments are
now correct. Make the tests supply the correct day-of-week to OSWORD &0F
so the Master has a chance of passing them; I don't know if it will or
not.

I think this still tests the IBOS implementation fine - if it ignores
the correct supplied day-of-week and calculates an incorrect day-of-week
to use instead we'll notice. We aren't testing that it *is* ignoring the
supplied day-of-week, but that's not particularly important.