Fix broken full reset
I had moved ZeroPageAUpToC0 in order to be able to re-use its final RTS
(to save a byte), but I didn't realise this was part of the code copied
into main RAM for a full reset.

This commit moves ZeroPageAUpToC0 back to its former location and
instead moves SetTransientCmdPtr so its RTS can be re-used.